+++ forui/lib/src/widgets/line_calendar/calendar_layout.dart
@@ -11,16 +11,20 @@ import 'package:forui/src/widgets/line_calendar/line_calendar_item.dart';
 
 @internal
 class CalendarLayout extends StatefulWidget {
-  final FCalendarController<DateTime?> controller;
+  final FCalendarController<DateTime?>? controller;
   final FLineCalendarStyle style;
   final AlignmentDirectional alignment;
+  final ScrollPhysics? physics;
   final double? cacheExtent;
+  final ScrollViewKeyboardDismissBehavior keyboardDismissBehavior;
   final TextScaler scale;
   final TextStyle textStyle;
   final ValueWidgetBuilder<FLineCalendarItemData> builder;
+  final ValueChanged<DateTime?>? onChange;
   final LocalDate start;
   final LocalDate? end;
-  final LocalDate? initial;
+  final LocalDate? initialScroll;
+  final LocalDate? initialSelection;
   final LocalDate today;
   final BoxConstraints constraints;
 
@@ -28,13 +32,17 @@ class CalendarLayout extends StatefulWidget {
     required this.controller,
     required this.style,
     required this.alignment,
+    required this.physics,
     required this.cacheExtent,
+    required this.keyboardDismissBehavior,
     required this.scale,
     required this.textStyle,
     required this.builder,
+    required this.onChange,
     required this.start,
     required this.end,
-    required this.initial,
+    required this.initialScroll,
+    required this.initialSelection,
     required this.today,
     required this.constraints,
     super.key,
@@ -50,19 +58,24 @@ class CalendarLayout extends StatefulWidget {
       ..add(DiagnosticsProperty('controller', controller))
       ..add(DiagnosticsProperty('style', style))
       ..add(DiagnosticsProperty('alignment', alignment))
+      ..add(DiagnosticsProperty('physics', physics))
       ..add(DoubleProperty('cacheExtent', cacheExtent))
+      ..add(DiagnosticsProperty('keyboardDismissBehavior', keyboardDismissBehavior))
       ..add(DiagnosticsProperty('scaler', scale))
       ..add(DiagnosticsProperty('textStyle', textStyle))
       ..add(ObjectFlagProperty.has('builder', builder))
+      ..add(ObjectFlagProperty.has('onChange', onChange))
       ..add(DiagnosticsProperty('start', start))
       ..add(DiagnosticsProperty('end', end))
-      ..add(DiagnosticsProperty('initial', initial))
+      ..add(DiagnosticsProperty('initialScroll', initialScroll))
+      ..add(DiagnosticsProperty('initialSelection', initialSelection))
       ..add(DiagnosticsProperty('today', today))
       ..add(DiagnosticsProperty('constraints', constraints));
   }
 }
 
 class _CalendarLayoutState extends State<CalendarLayout> {
+  late FCalendarController<DateTime?> _controller;
   late ScrollController _scrollController;
   late double _width;
 
@@ -71,7 +84,10 @@ class _CalendarLayoutState extends State<CalendarLayout> {
     super.initState();
     _width = _estimateWidth();
 
-    final start = ((widget.initial ?? widget.today).difference(widget.start).inDays) * _width;
+    _controller = widget.controller ?? FCalendarController.date(initialSelection: widget.initialSelection?.toNative());
+    _controller.addValueListener(_onChange);
+
+    final start = ((widget.initialScroll ?? widget.today).difference(widget.start).inDays) * _width;
     _scrollController = ScrollController(
       initialScrollOffset: switch (widget.alignment.start) {
         -1 => start,
@@ -86,6 +102,17 @@ class _CalendarLayoutState extends State<CalendarLayout> {
     super.didUpdateWidget(old);
     if (widget.style != old.style || widget.scale != old.scale || widget.textStyle != old.textStyle) {
       _width = _estimateWidth();
+    }
+
+    if (widget.controller != old.controller) {
+      if (old.controller == null) {
+        _controller.dispose();
+      } else {
+        _controller.removeValueListener(_onChange);
+      }
+
+      _controller = widget.controller ?? FCalendarController.date(initialSelection: _controller.value);
+      _controller.addValueListener(_onChange);
     }
   }
 
@@ -110,6 +137,8 @@ class _CalendarLayoutState extends State<CalendarLayout> {
     ].max!;
   }
 
+  void _onChange(DateTime? date) => widget.onChange?.call(date);
+
   @override
   Widget build(BuildContext _) {
     final placeholder = widget.today.toNative();
@@ -123,7 +152,9 @@ class _CalendarLayoutState extends State<CalendarLayout> {
           controller: _scrollController,
           scrollDirection: Axis.horizontal,
           padding: EdgeInsets.zero,
+          physics: widget.physics,
           cacheExtent: widget.cacheExtent,
+          keyboardDismissBehavior: widget.keyboardDismissBehavior,
           itemExtent: _width,
           itemCount: widget.end == null ? null : widget.end!.difference(widget.start).inDays + 1,
           itemBuilder: (_, index) {
@@ -131,7 +162,7 @@ class _CalendarLayoutState extends State<CalendarLayout> {
             return Padding(
               padding: widget.style.padding,
               child: Item(
-                controller: widget.controller,
+                controller: _controller,
                 style: widget.style,
                 date: date.toNative(),
                 today: widget.today == date,
@@ -147,6 +178,11 @@ class _CalendarLayoutState extends State<CalendarLayout> {
   @override
   void dispose() {
     _scrollController.dispose();
+    if (widget.controller == null) {
+      _controller.dispose();
+    } else {
+      _controller.removeValueListener(_onChange);
+    }
     super.dispose();
   }
 }